refactor: 의존성 흐름 위반 해결
